requirements.txt
black==24.10.0
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
Pygments==2.18.0
PyYAML==6.0.2
rich==13.9.4
shellingham==1.5.4
typer==0.13.0
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
